starting modifications for wav stimulus curve
no good way to deal with playing multiple wav stimuli, so created
standalone curve type. the Run Curve function in opto will now check for
standalone stimulus type. If this is true, the script in test.Script
will be executed. This script will need to do all the setup of stimuli
and hardware on its own.

Starting on this. next is to work on an example, derived from
opto_playCache function